Merge 3.0.9-WEBC into 3.0.9

+++ KalturaWowzaServer/src/com/kaltura/media/server/wowza/RecordingManager.java
@@ -47,7 +47,7 @@ public class RecordingManager {
 		private String entryId;
 		private String assetId;
 		private KalturaMediaServerIndex index;
-		private float appendedDuration = 0;
+		private double appendedDuration = 0;
 		
 		public EntryRecorder(String entryId, String assetId, KalturaMediaServerIndex index) {
 			super();
@@ -85,8 +85,8 @@ public class RecordingManager {
 			
 			logger.info("Stream [" + stream.getName() + "] file [" + file.getAbsolutePath() + "] folder [" + file.getParent() + "]");
 			Date date = new Date();
-			float duration = ((float)(date.getTime() - liveStreamRecord.getStartTime().getMillis())) / 1000;
-			float currentChunkDuration = duration - appendedDuration;
+			double duration = ((double)(date.getTime() - liveStreamRecord.getStartTime().getMillis())) / 1000;
+			double currentChunkDuration = duration - appendedDuration;
 			appendedDuration = duration;
 
 			if(group != null){
